LayoutTests/crypto/subtle/ecdh-import-pkcs8-key-p256-expected.txt

Test importing a P-256 PKCS8 ECDH key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'private'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDH'
PASS publicKey.algorithm.namedCurve is 'P-256'
PASS publicKey.usages is [ 'deriveBits', 'deriveKey' ]
PASS successfullyParsed is true

TEST COMPLETE
